# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VBUY — Premium Ghanaian E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order (guest or authenticated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID (guest via token or authenticated user)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
  /orders/{id}/momo-reference:
    post:
      operationId: submitMomoReference
      tags: [orders]
      summary: Submit Mobile Money transaction reference for manual verification
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MomoReferenceInput"
      responses:
        "200":
          description: Reference recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MomoReferenceResponse"
        "400":
          description: Validation error
        "403":
          description: Forbidden
        "404":
          description: Order not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      properties:
        id:             { type: integer }
        digitalAddress: { type: string }
        region:         { type: string }
        district:       { type: string }
        notes:          { type: ["string", "null"] }
        userId:         { type: ["string", "null"] }
      required: [id, digitalAddress, region, district]
    AddressInput:
      type: object
      properties:
        digitalAddress: { type: string }
        region:         { type: string }
        district:       { type: string }
        notes:          { type: string }
        userId:         { type: string }
      required: [digitalAddress, region, district]
    OrderItem:
      type: object
      properties:
        id:               { type: integer }
        productVariantId: { type: integer }
        quantity:         { type: integer }
        unitPrice:        { type: number }
        productName:      { type: string }
        productImage:     { type: ["string", "null"] }
        variantName:      { type: string }
      required: [id, productVariantId, quantity, unitPrice, productName, variantName]
    Order:
      type: object
      properties:
        id:            { type: integer }
        status:        { type: string }
        paymentMethod: { type: string }
        paymentStatus: { type: string }
        total:         { type: number }
        deliveryFee:   { type: number }
        deliveryRegion:
          type: ["string", "null"]
          description: Snapshot zone — accra | outside_accra
        userId:        { type: ["string", "null"] }
        carrier:       { type: ["string", "null"] }
        trackingNumber: { type: ["string", "null"] }
        trackingUrl:   { type: ["string", "null"] }
        momoReference: { type: ["string", "null"] }
        paymentSubmittedAt: { type: ["string", "null"] }
        paymentVerifiedAt: { type: ["string", "null"] }
        createdAt:     { type: string }
        address:
          $ref: "#/components/schemas/Address"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [id, status, paymentMethod, paymentStatus, total, deliveryFee, createdAt, items]
    OrderInput:
      type: object
      properties:
        sessionId:
          type: string
        address:
          $ref: "#/components/schemas/AddressInput"
        paymentMethod:
          type: string
          enum: [momo_manual]
        deliveryZone:
          type: string
          enum: [accra, outside_accra]
          description: Flat delivery fee zone selected at checkout
        userId:
          type: string
        turnstileToken:
          type: string
          minLength: 1
          description: Cloudflare Turnstile token (required when TURNSTILE_SECRET_KEY is set)
      required: [sessionId, address, paymentMethod, deliveryZone, turnstileToken]
    MomoReferenceInput:
      type: object
      properties:
        momoReference: { type: string, minLength: 3 }
        turnstileToken:
          type: string
          minLength: 1
          description: Cloudflare Turnstile token (required when TURNSTILE_SECRET_KEY is set)
      required: [momoReference, turnstileToken]
    MomoReferenceResponse:
      type: object
      properties:
        order:
          $ref: "#/components/schemas/Order"
        duplicateReference: { type: boolean }
      required: [order, duplicateReference]
...